Added 'My Libraries' feature for users to view and manage their own libraries, with separate views and permissions.

+++ controllers/libraries.js
@@ -5,11 +5,17 @@ const multer = require('../config/multer') // Multer for image upload
 const isSignedIn = require('../middleware/is-signed-in')
 
 //GET
-//Show all libraries
+//Show all libraries(View Only)
 router.get('/', async (req, res) => {
   const libraries = await Library.find().populate('userId')
-  res.render('libraries/index.ejs', { libraries })
-})
+ res.render('libraries/index.ejs', { libraries, user: req.session.user })
+});
+// Show user libraries (Editable)
+router.get('/my-libraries', isSignedIn, async (req, res) => {
+  const libraries = await Library.find({ userId: req.session.user._id }).populate('userId');
+  res.render('libraries/my-libraries.ejs', { libraries, user: req.session.user });
+});
+
 //Form to create new library
 router.get('/new', isSignedIn, (req, res) => {
   res.render('libraries/new.ejs')
@@ -31,18 +37,22 @@ router.post('/', isSignedIn, multer.single('image'), async (req, res) => {
     userId: req.session.user._id
   })
   await newLibrary.save()
-  res.redirect('/libraries')
+  res.redirect('/libraries/my-libraries')
 })
 
 //Show library details
 router.get('/:id', async (req, res) => {
   const library = await Library.findById(req.params.id).populate('userId')
-  res.render('libraries/show.ejs', { library })
+  res.render('libraries/show.ejs', { library, user: req.session.user })
 })
+
 
 // Edit form
 router.get('/:id/edit', isSignedIn, async (req, res) => {
   const library = await Library.findById(req.params.id)
+    if (!library || !library.userId.equals(req.session.user._id)) {
+    return res.redirect('/libraries');
+  }
   res.render('libraries/edit.ejs', { library })
 })
 
@@ -50,6 +60,10 @@ router.get('/:id/edit', isSignedIn, async (req, res) => {
 //  Update library
 router.put('/:id', isSignedIn, multer.single('image'), async (req, res) => {
   const library = await Library.findById(req.params.id)
+    if (!library || !library.userId.equals(req.session.user._id)) {
+    return res.redirect('/libraries');
+  }
+
   library.name = req.body.name
   library.location = req.body.location
   library.address = req.body.address
@@ -65,8 +79,12 @@ router.put('/:id', isSignedIn, multer.single('image'), async (req, res) => {
 
 //Delete library
 router.delete('/:id', isSignedIn, async (req, res) => {
-  await Library.findByIdAndDelete(req.params.id)
-  res.redirect('/libraries')
-})
-
+  const library = await Library.findById(req.params.id);
+  if (!library || !library.userId.equals(req.session.user._id)) {
+    return res.redirect('/libraries');
+  }
+  
+  await Library.findByIdAndDelete(req.params.id);
+  res.redirect('/libraries/my-libraries');
+});
 module.exports = router
